Add mock test.
